Allow require network with a timeout.
This CL introduces a new Params API where the network requirement can be
set to timeout. It is useful if you want to run some requiest or inform
the user if you cannot after a certain time.
This is also a preliminary step to introduce WIFI requirement since it
is useless without a timeout (there are many devices which will never see
WIFI